+++ packages/callkit-uniapp-mp-weixin/App.vue
@@ -158,7 +158,7 @@ function startAudioCall() {
     return
   }
   uni.navigateTo({
-    url: `/pages/meeting/meeting?targetUserId=${targetUserId.value}&callType=audio`
+    url: `/uni_modules/easemob-callkit-mp-weixin/pages/single-call-page/single-call-page?targetUserId=${targetUserId.value}&callType=audio`
   })
 }
 
@@ -168,7 +168,7 @@ function startVideoCall() {
     return
   }
   uni.navigateTo({
-    url: `/pages/meeting/meeting?targetUserId=${targetUserId.value}&callType=video`
+    url: `/uni_modules/easemob-callkit-mp-weixin/pages/single-call-page/single-call-page?targetUserId=${targetUserId.value}&callType=video`
   })
 }
 </script>